Move RoadmapGraphSidebar to VersionDetailsModal
This accomplishes navigation from VersionDetailsModal to task overview
and back.
